Add dash to dock and blur my shell to fedora gnome

+++ Bazzite/Setup.sh
@@ -29,8 +29,7 @@ install_ghostty() {
     sudo mkdir -p /etc/yum.repos.d
     curl -fsSL "https://copr.fedorainfracloud.org/coprs/scottames/ghostty/repo/fedora-${VERSION_ID}/scottames-ghostty-fedora-${VERSION_ID}.repo" \
     | sudo tee /etc/yum.repos.d/_copr:copr.fedorainfracloud.org:scottames:ghostty.repo >/dev/null
-    
-    sudo rpm-ostree refresh-md
+
     sudo rpm-ostree install ghostty
 }
 
@@ -60,27 +59,36 @@ install_fonts() {
     local fonts_dir="$HOME/.local/share/fonts"
     local jb_zip="$fonts_dir/JetBrainsMono.zip"
     local jb_tmp_dir="$fonts_dir/jetbrainsmono_nf_tmp"
+    local fonts_updated=0
     
     mkdir -p "$fonts_dir"
     
-    # Instal MesloLGS fonts
-    curl -fLo "$fonts_dir/MesloLGS NF Regular.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Regular.ttf
-    curl -fLo "$fonts_dir/MesloLGS NF Bold.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Bold.ttf
-    curl -fLo "$fonts_dir/MesloLGS NF Italic.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Italic.ttf
-    curl -fLo "$fonts_dir/MesloLGS NF Bold Italic.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Bold%20Italic.ttf
+    if ! fc-list | grep -Eiq 'MesloLGS (NF|Nerd Font)'; then
+        # Install MesloLGS fonts
+        curl -fLo "$fonts_dir/MesloLGS NF Regular.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Regular.ttf
+        curl -fLo "$fonts_dir/MesloLGS NF Bold.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Bold.ttf
+        curl -fLo "$fonts_dir/MesloLGS NF Italic.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Italic.ttf
+        curl -fLo "$fonts_dir/MesloLGS NF Bold Italic.ttf" https://github.com/romkatv/powerlevel10k-media/raw/master/MesloLGS%20NF%20Bold%20Italic.ttf
+        fonts_updated=1
+    fi
     
-    # Install jetbrain fonts
-    curl -fLo "$jb_zip" https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip
+    if ! fc-list | grep -Eiq 'JetBrainsMono.*(NF|Nerd Font)'; then
+        # Install JetBrains Mono Nerd Font
+        curl -fLo "$jb_zip" https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip
     
-    rm -rf "$jb_tmp_dir"
-    mkdir -p "$jb_tmp_dir"
+        rm -rf "$jb_tmp_dir"
+        mkdir -p "$jb_tmp_dir"
     
-    unzip -o "$jb_zip" -d "$jb_tmp_dir"
-    find "$jb_tmp_dir" -type f \( -name "*.ttf" -o -name "*.otf" \) -exec cp {} "$fonts_dir/" \;
+        unzip -o "$jb_zip" -d "$jb_tmp_dir"
+        find "$jb_tmp_dir" -type f \( -name "*.ttf" -o -name "*.otf" \) -exec cp {} "$fonts_dir/" \;
     
-    rm -rf "$jb_tmp_dir" "$jb_zip"
+        rm -rf "$jb_tmp_dir" "$jb_zip"
+        fonts_updated=1
+    fi
 
-    fc-cache -f "$fonts_dir"
+    if (( fonts_updated )); then
+        fc-cache -f "$fonts_dir"
+    fi
 }
 
 main() {
